+++ Code/ShipWorks/Data/Import/Xml/GenericXmlOrderLoader.cs
@@ -8,6 +8,7 @@ using Interapptive.Shared.Utility;
 using ShipWorks.Stores.Content;
 using Interapptive.Shared.Business;
 using Interapptive.Shared.Business.Geography;
+using Interapptive.Shared.Enums;
 using ShipWorks.Stores.Communication;
 using ShipWorks.Stores.Platforms.Amazon;
 
@@ -91,7 +92,7 @@ namespace ShipWorks.Data.Import.Xml
         private static void LoadAmazonOrderDetails(GenericModuleOrderEntity order, XPathNavigator xpath)
         {
             bool isPrime = XPathUtility.Evaluate(xpath, "Amazon/IsPrime", false);
-            order.IsPrime = isPrime ? (int) AmazonIsPrime.Yes : (int) AmazonIsPrime.No;
+            order.IsPrime = isPrime ? AmazonIsPrime.Yes : AmazonIsPrime.No;
             order.IsFBA = XPathUtility.Evaluate(xpath, "Amazon/IsFBA", false);
             order.AmazonOrderID = XPathUtility.Evaluate(xpath, "Amazon/AmazonOrderID", string.Empty);
             order.IsSameDay = XPathUtility.Evaluate(xpath, "Amazon/IsSameDay", false);
